Python/requirements.txt
python-dotenv
redis